Time: 4 ms (26.33%), Space: 44.8 MB (59.84%) - LeetHub

+++ 0713-subarray-product-less-than-k/0713-subarray-product-less-than-k.cpp
@@ -8,7 +8,7 @@ public:
         int r=0;
 
         
-        long long mul=1;
+        int  mul=1;
         while(r<n){
            
              mul=mul*nums[r];
